Conflicts:
	app/controllers/user/Rating.java

+++ app/controllers/user/Rating.java
@@ -3,6 +3,7 @@ package controllers.user;
 import controllers.Stuff;
 import models.Account;
 import models.Criteria;
+import models.Setting;
 import models.Team;
 import play.mvc.Controller;
 import play.mvc.Result;
@@ -18,56 +19,55 @@ import java.util.*;
 public class Rating extends Controller{
     @Security.Authenticated(Secured.class)
     public static Result ratingResult() {
-        List<Stuff> rankAll = new ArrayList();
-        List<Criteria> cri = Criteria.getall();
-        List<Team> team = Team.getAll();
-        double[] overAll = new double[team.size()];
+        if (Setting.find.byId(Setting.RATING_RESULT).isActivated) {
+            List<Stuff> rankAll = new ArrayList();
+            List<Criteria> cri = Criteria.getall();
+            List<Team> team = Team.getAll();
+            double[] overAll = new double[team.size()];
 
-        for(int i=0; i<cri.size(); i++) {
-            HashMap<Team, Double> rankInCri = new HashMap<>();
-            for(int j=0; j<team.size(); j++) {
-                List<models.Rating> data = models.Rating.getRatingSpecific(cri.get(i), team.get(j));
-                double scoreAvg = 0;
-                for (int k = 0; k<data.size(); k++) {
-                    scoreAvg += data.get(k).rating;
+            for (int i = 0; i < cri.size(); i++) {
+                HashMap<Team, Double> rankInCri = new HashMap<>();
+                for (int j = 0; j < team.size(); j++) {
+                    List<models.Rating> data = models.Rating.getRatingSpecific(cri.get(i), team.get(j));
+                    double scoreAvg = 0;
+                    for (int k = 0; k < data.size(); k++) {
+                        scoreAvg += data.get(k).rating;
+                    }
+                    scoreAvg = scoreAvg / data.size();
+                    //-----------------------------------------------//
+                    overAll[j] += scoreAvg;
+                    //-----------------------------------------------//
+                    rankInCri.put(team.get(j), scoreAvg);
                 }
-                scoreAvg = scoreAvg / data.size();
-                //-----------------------------------------------//
-                overAll[j] += scoreAvg;
-                //-----------------------------------------------//
-                rankInCri.put(team.get(j), scoreAvg);
+                Set<Map.Entry<Team, Double>> set = rankInCri.entrySet();
+                List<Map.Entry<Team, Double>> list = new ArrayList<Map.Entry<Team, Double>>(set);
+                Collections.sort(list, new Comparator<Map.Entry<Team, Double>>() {
+                    public int compare(Map.Entry<Team, Double> o1, Map.Entry<Team, Double> o2) {
+                        return (o2.getValue()).compareTo(o1.getValue());
+                    }
+                });
+                Stuff stuff = new Stuff(list, cri.get(i));
+                rankAll.add(stuff);
             }
-            Set<Map.Entry<Team, Double>> set = rankInCri.entrySet();
+            for (int i = 0; i < overAll.length; i++) {
+                overAll[i] /= cri.size();
+            }
+            //-------------------------------------------------//
+            HashMap<Team, Double> rankOverAll = new HashMap<>();
+            for (int i = 0; i < team.size(); i++) {
+                rankOverAll.put(team.get(i), overAll[i]);
+            }
+            Set<Map.Entry<Team, Double>> set = rankOverAll.entrySet();
             List<Map.Entry<Team, Double>> list = new ArrayList<Map.Entry<Team, Double>>(set);
-            Collections.sort( list, new Comparator<Map.Entry<Team, Double>>()
-            {
-                public int compare( Map.Entry<Team, Double> o1, Map.Entry<Team, Double> o2 )
-                {
-                    return (o2.getValue()).compareTo( o1.getValue() );
+            Collections.sort(list, new Comparator<Map.Entry<Team, Double>>() {
+                public int compare(Map.Entry<Team, Double> o1, Map.Entry<Team, Double> o2) {
+                    return (o2.getValue()).compareTo(o1.getValue());
                 }
-            } );
-            Stuff stuff = new Stuff(list, cri.get(i));
-            rankAll.add(stuff);
+            });
+            //-------------------------------------------------//
+            return ok(rating_result.render(rankAll, list));
         }
-        for (int i=0;i<overAll.length;i++){
-            overAll[i] /= cri.size();
-        }
-        //-------------------------------------------------//
-        HashMap<Team, Double> rankOverAll = new HashMap<>();
-        for(int i=0; i<team.size(); i++) {
-            rankOverAll.put(team.get(i), overAll[i]);
-        }
-        Set<Map.Entry<Team, Double>> set = rankOverAll.entrySet();
-        List<Map.Entry<Team, Double>> list = new ArrayList<Map.Entry<Team, Double>>(set);
-        Collections.sort( list, new Comparator<Map.Entry<Team, Double>>()
-        {
-            public int compare( Map.Entry<Team, Double> o1, Map.Entry<Team, Double> o2 )
-            {
-                return (o2.getValue()).compareTo( o1.getValue() );
-            }
-        } );
-        //-------------------------------------------------//
-        return ok(rating_result.render(rankAll, list));
+        return badRequest("Disable this function by admin");
     }
 
     @Security.Authenticated(Secured.class)
@@ -90,29 +90,35 @@ public class Rating extends Controller{
         j++;
         }
         return result;
+        if (Setting.find.byId(Setting.RATING).isActivated) {
+            return ok(rating.render(Criteria.getall(), teamID));
+        }
+        return badRequest("Disable this function by admin");
     }
 
     @Security.Authenticated(Secured.class)
     public static Result ratingPost() {
-        if (request().method().equals("POST")) {
-            Team team = null;
-            Account account = Account.findEmail(session().get("email"));
-            Map<String, String[]> map = request().body().asFormUrlEncoded();
-            Long id = Long.parseLong(map.get("uid")[0]);
-            team = Team.find.byId(id);
-            for(models.Rating rating: models.Rating.findByTeamAndAccount(id, account.id)){
-                rating.delete();
-            }
-            for (Map.Entry<String, String[]> entry : map.entrySet()) {
-                String key = entry.getKey();
-                String[] value = entry.getValue();
-                if (key.equals("uid")) break;
-                models.Rating obj = new models.Rating(account, Criteria.find.byId(Long.parseLong(key)), Integer.parseInt(value[0]), team);
-                obj.save();
-            }
-            return ok();
+        if (Setting.find.byId(Setting.RATING).isActivated) {
+            if (request().method().equals("POST")) {
+                Team team = null;
+                Account account = Account.findEmail(session().get("email"));
+                Map<String, String[]> map = request().body().asFormUrlEncoded();
+                Long id = Long.parseLong(map.get("uid")[0]);
+                team = Team.find.byId(id);
+                for (models.Rating rating : models.Rating.findByTeamAndAccount(id, account.id)) {
+                    rating.delete();
+                }
+                for (Map.Entry<String, String[]> entry : map.entrySet()) {
+                    String key = entry.getKey();
+                    String[] value = entry.getValue();
+                    if (key.equals("uid")) break;
+                    models.Rating obj = new models.Rating(account, Criteria.find.byId(Long.parseLong(key)), Integer.parseInt(value[0]), team);
+                    obj.save();
+                }
+                return ok();
+            } else return ok();
         }
-        else return ok();
+        return badRequest("Disable this function by admin");
     }
 
 }
